+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/io/benchmark/StreamNetworkPointToPointBenchmark.java
@@ -19,7 +19,6 @@
 package org.apache.flink.streaming.runtime.io.benchmark;
 
 import org.apache.flink.runtime.io.network.api.writer.RecordWriter;
-import org.apache.flink.runtime.io.network.benchmark.ReceiverThread;
 import org.apache.flink.types.LongValue;
 
 import java.util.concurrent.CompletableFuture;
@@ -74,7 +73,7 @@ public class StreamNetworkPointToPointBenchmark {
 		environment.setUp(1, 1);
 
 		receiver = environment.createReceiver();
-		recordWriter = environment.createStreamRecordWriter(0, flushTimeout);
+		recordWriter = environment.createRecordWriter(0, flushTimeout);
 	}
 
 	/**
